TST/REF: collect tests by method, some misplaced
